Tasks-Management/backend/requirements.txt
djangorestframework==3.14
django-cors-headers==4.1.0
djoser==2.2.0
loguru==0.7.0
python-dotenv==1.0
djangorestframework_simplejwt==5.3.0
